central-api/requirements.txt
FastAPI
uvicorn
paho-mqtt
pydantic
httpx